Rename routes to tabBarRoutes
