+++ src/features/internal/components/DeeplinksGeneratorForm.native.test.tsx
@@ -14,7 +14,7 @@ jest.mock('libs/subcategories/useSubcategories')
 mockdate.set(new Date('2022-08-09T00:00:00Z'))
 
 jest.mock('libs/firebase/analytics/analytics')
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 
 jest.mock('react-native/Libraries/Animated/createAnimatedComponent', () => {
   return function createAnimatedComponent(Component: unknown) {
