+++ src/features/internal/components/DeeplinksGeneratorForm.tsx
@@ -25,7 +25,7 @@ import { getScreenPath } from 'features/navigation/RootNavigator/linking/getScre
 import { isSearchStackScreen } from 'features/navigation/SearchStackNavigator/searchRoutes'
 import { getSearchStackConfig } from 'features/navigation/SearchStackNavigator/searchStackHelpers'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
-import { isTabScreen } from 'features/navigation/TabBar/routes'
+import { isTabScreen } from 'features/navigation/TabBar/tabBarRoutes'
 import { MAX_PRICE_IN_CENTS } from 'features/search/helpers/reducer.helpers'
 import { LocationFilter } from 'features/search/types'
 import { env } from 'libs/environment/env'
