+++ src/features/navigation/RootNavigator/__mocks__/routes.tsx
@@ -1,5 +1,5 @@
 import { screenParamsParser } from 'features/navigation/screenParamsUtils'
-import { tabNavigatorPathConfig } from 'features/navigation/TabBar/__mocks__/routes'
+import { tabNavigatorPathConfig } from 'features/navigation/TabBar/__mocks__/tabBarRoutes'
 
 import { RootRoute } from '../types'
