+++ src/features/navigation/TabBar/TabNavigator.tsx
@@ -1,8 +1,8 @@
 import { BottomTabBarProps, BottomTabNavigationOptions } from '@react-navigation/bottom-tabs'
 import React from 'react'
 
-import { initialRouteName, TabScreens } from 'features/navigation/TabBar/routes'
 import { TabStack } from 'features/navigation/TabBar/Stack'
+import { initialRouteName, TabScreens } from 'features/navigation/TabBar/tabBarRoutes'
 
 import { TabBar } from './TabBar'
