+++ src/features/navigation/TabBar/tabBarRoutes.ts
@@ -18,7 +18,7 @@ export const initialRouteName = 'Home'
 
 const Home = withAsyncErrorBoundary(HomeComponent)
 
-const routes: TabRoute[] = [
+const tabBarRoutes: TabRoute[] = [
   {
     name: 'Home',
     component: Home,
@@ -52,12 +52,12 @@ const routes: TabRoute[] = [
 ]
 
 export function isTabScreen(screen: string): screen is TabRouteName {
-  const tabRouteNames = routes.map((route): string => route.name)
+  const tabRouteNames = tabBarRoutes.map((route): string => route.name)
   return tabRouteNames.includes(screen)
 }
 
 const { screensConfig: tabScreensConfig, Screens: TabScreens } = getScreensAndConfig(
-  routes,
+  tabBarRoutes,
   TabStack.Screen
 )
