+++ src/features/profile/pages/NewEmailSelection/NewEmailSelection.web.test.tsx
@@ -12,7 +12,7 @@ jest.mock('uuid', () => {
     v4: jest.fn(() => value++),
   }
 })
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 
 describe('<NewEmailSelection />', () => {
   describe('Accessibility', () => {
