+++ src/features/search/components/AutocompleteOffer/AutocompleteOffer.native.test.tsx
@@ -28,7 +28,7 @@ jest.mock('features/search/context/SearchWrapper', () => ({
 }))
 
 jest.mock('libs/firebase/analytics/analytics')
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 
 describe('AutocompleteOffer component', () => {
   beforeEach(() => {
