+++ src/features/search/components/CategoriesList/CategoriesList.native.test.tsx
@@ -51,7 +51,7 @@ jest.mock('libs/location', () => ({
 }))
 
 jest.mock('libs/firebase/analytics/analytics')
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 
 const user = userEvent.setup()
