+++ src/features/search/components/SearchHeader/SearchHeader.native.test.tsx
@@ -52,7 +52,7 @@ jest.spyOn(useFilterCountAPI, 'useFilterCount').mockReturnValue(3)
 jest.useFakeTimers()
 
 jest.mock('libs/firebase/analytics/analytics')
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 
 describe('SearchHeader component', () => {
   beforeEach(() => {
