+++ src/features/search/components/SearchHeader/SearchHeader.web.test.tsx
@@ -41,7 +41,7 @@ jest.mock('features/search/context/SearchWrapper', () => ({
 jest.mock('libs/firebase/analytics/analytics')
 jest.mock('libs/firebase/remoteConfig/remoteConfig.services')
 jest.spyOn(useRemoteConfigQuery, 'useRemoteConfigQuery').mockReturnValue(DEFAULT_REMOTE_CONFIG)
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 
 describe('SearchHeader component', () => {
   beforeEach(() => {
