+++ src/features/search/pages/SearchFilter/SearchFilter.native.test.tsx
@@ -57,7 +57,7 @@ jest.mock('features/search/api/useSearchResults/useSearchResults', () => ({
 }))
 
 jest.mock('libs/firebase/analytics/analytics')
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 
 jest.mock('react-native/Libraries/Animated/createAnimatedComponent', () => {
   return function createAnimatedComponent(Component: unknown) {
