+++ src/features/search/pages/SearchFilter/SearchFilter.web.test.tsx
@@ -45,7 +45,7 @@ jest.mock('features/search/context/SearchWrapper', () => ({
 
 jest.mock('libs/firebase/analytics/analytics')
 jest.mock('libs/firebase/remoteConfig/remoteConfig.services')
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 
 describe('<SearchFilter/>', () => {
   beforeEach(() => {
