+++ src/features/search/pages/SearchLanding/SearchLanding.native.test.tsx
@@ -213,7 +213,7 @@ const mockedEmptyHistory = {
 }
 
 jest.mock('libs/firebase/analytics/analytics')
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 
 jest.mock('react-native/Libraries/Animated/createAnimatedComponent', () => {
   return function createAnimatedComponent(Component: unknown) {
