+++ src/features/search/pages/SearchLanding/SearchLanding.web.test.tsx
@@ -79,7 +79,7 @@ jest.mock('uuid', () => ({
 
 jest.mock('libs/firebase/analytics/analytics')
 jest.mock('libs/firebase/remoteConfig/remoteConfig.services')
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 jest.mock('ui/theme/customFocusOutline/customFocusOutline')
 
 describe('<SearchLanding />', () => {
