+++ src/features/search/pages/ThematicSearch/ThematicSearch.web.test.tsx
@@ -16,7 +16,7 @@ import { checkAccessibilityFor, render, screen } from 'tests/utils/web'
 
 jest.mock('libs/firebase/analytics/analytics')
 jest.mock('libs/firebase/remoteConfig/remoteConfig.services')
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 jest.mock('ui/theme/customFocusOutline/customFocusOutline')
 
 const mockDispatch = jest.fn()
