+++ src/ui/components/buttons/SubcategoryButton/SubcategoryButton.native.test.tsx
@@ -10,7 +10,7 @@ import { theme } from 'theme'
 import { SubcategoryButton } from 'ui/components/buttons/SubcategoryButton/SubcategoryButton'
 
 jest.mock('libs/firebase/analytics/analytics')
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 
 const mockSearchState = initialSearchState
 const mockDispatch = jest.fn()
