+++ src/ui/components/buttons/SubcategoryButton/SubcategoryButtonList.native.test.tsx
@@ -35,7 +35,7 @@ jest.mock('features/search/context/SearchWrapper', () => ({
 }))
 
 jest.mock('libs/firebase/analytics/analytics')
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 
 describe('<SubcategoryButtonList/>', () => {
   it('should render SubcategoryButtonList', async () => {
