+++ src/ui/components/buttons/SubcategoryButton/SubcategoryButtonListWrapper.native.test.tsx
@@ -17,7 +17,7 @@ jest.mock('features/search/context/SearchWrapper', () => ({
 }))
 
 jest.mock('libs/firebase/analytics/analytics')
-jest.mock('features/navigation/TabBar/routes')
+jest.mock('features/navigation/TabBar/tabBarRoutes')
 
 const user = userEvent.setup()
 jest.useFakeTimers()
